+++ packages/core/rust/index.js
@@ -310,7 +310,7 @@ if (!nativeBinding) {
   throw new Error(`Failed to load native binding`)
 }
 
-const { initTracingSubscriber, Lmdb, findAncestorFile, findFirstFile, findNodeModule, hashString, hashBuffer, Hash, optimizeImage, AtlaspackTracer, createAssetId, atlaspackNapiCreate, atlaspackNapiBuildAssetGraph, atlaspackNapiRespondToFsEvents, createDependencyId, createEnvironmentId, setAllEnvironments, getAllEnvironments, getEnvironment, addEnvironment, getAvailableThreads, initializeMonitoring, closeMonitoring, newNodejsWorker, runInlineRequiresOptimizer, runInlineRequiresOptimizerAsync, Resolver, performStringReplacements, transform, transformAsync, getVcsStateSnapshot, getEventsSince } = nativeBinding
+const { initTracingSubscriber, Lmdb, findAncestorFile, findFirstFile, findNodeModule, hashString, hashBuffer, Hash, optimizeImage, AtlaspackTracer, createAssetId, atlaspackNapiCreate, atlaspackNapiBuildAssetGraph, atlaspackNapiRespondToFsEvents, createDependencyId, createEnvironmentId, setAllEnvironments, getAllEnvironments, getEnvironment, addEnvironment, getAvailableThreads, initializeMonitoring, closeMonitoring, getNativeMemoryStats, resetMemoryTracking, sampleNativeMemory, newNodejsWorker, runInlineRequiresOptimizer, runInlineRequiresOptimizerAsync, applyCompiledCssInJsPlugin, Resolver, performStringReplacements, transform, transformAsync, getVcsStateSnapshot, getEventsSince, applyTokensPlugin } = nativeBinding
 
 module.exports.initTracingSubscriber = initTracingSubscriber
 module.exports.Lmdb = Lmdb
@@ -335,12 +335,17 @@ module.exports.addEnvironment = addEnvironment
 module.exports.getAvailableThreads = getAvailableThreads
 module.exports.initializeMonitoring = initializeMonitoring
 module.exports.closeMonitoring = closeMonitoring
+module.exports.getNativeMemoryStats = getNativeMemoryStats
+module.exports.resetMemoryTracking = resetMemoryTracking
+module.exports.sampleNativeMemory = sampleNativeMemory
 module.exports.newNodejsWorker = newNodejsWorker
 module.exports.runInlineRequiresOptimizer = runInlineRequiresOptimizer
 module.exports.runInlineRequiresOptimizerAsync = runInlineRequiresOptimizerAsync
+module.exports.applyCompiledCssInJsPlugin = applyCompiledCssInJsPlugin
 module.exports.Resolver = Resolver
 module.exports.performStringReplacements = performStringReplacements
 module.exports.transform = transform
 module.exports.transformAsync = transformAsync
 module.exports.getVcsStateSnapshot = getVcsStateSnapshot
 module.exports.getEventsSince = getEventsSince
+module.exports.applyTokensPlugin = applyTokensPlugin