+++ packages/transformers/js/src/JSTransformer.ts
@@ -4,7 +4,6 @@ import type {
   SourceLocation,
   FilePath,
   FileCreateInvalidation,
-  ConditionMeta,
 } from '@atlaspack/types';
 import type {SchemaEntity} from '@atlaspack/utils';
 import type {Diagnostic} from '@atlaspack/diagnostic';
@@ -173,7 +172,7 @@ type MacroContext = {
 };
 
 export default new Transformer({
-  async loadConfig({config, options}) {
+  async loadConfig({config, options, logger}) {
     let packageJson = await config.getPackage();
     let isJSX,
       pragma,
@@ -321,8 +320,9 @@ export default new Transformer({
         CONFIG_SCHEMA,
         {
           data: conf.contents,
-          // FIXME
-          source: await options.inputFS.readFile(conf.filePath, 'utf8'),
+          source: getFeatureFlag('schemaValidationDeferSourceLoading')
+            ? () => options.inputFS.readFileSync(conf.filePath, 'utf8')
+            : await options.inputFS.readFile(conf.filePath, 'utf8'),
           filePath: conf.filePath,
           prependKey: `/${encodeJSONKeyComponent('@atlaspack/transformer-js')}`,
         },
@@ -541,10 +541,12 @@ export default new Transformer({
       magic_comments:
         Boolean(config?.magicComments) ||
         getFeatureFlag('supportWebpackChunkName'),
+      is_source: asset.isSource,
       enable_global_this_aliaser: Boolean(config.enableGlobalThisAliaser),
       enable_lazy_loading_transformer: Boolean(
         config.enableLazyLoadingTransformer,
       ),
+      nested_promise_import_fix: options.featureFlags.nestedPromiseImportFix,
       enable_dead_returns_remover: Boolean(config.enableDeadReturnsRemover),
       enable_unused_bindings_remover: Boolean(
         config.enableUnusedBindingsRemover,